Move vigencias to VigenciasInterface, use VigenciasTrait
