+++ src/EntryInterface.php
@@ -4,13 +4,9 @@ declare(strict_types=1);
 
 namespace PhpCfdi\SatCatalogos;
 
-interface EntryInterface
+interface EntryInterface extends VigenciasInterface
 {
     public function id(): string;
 
     public function texto(): string;
-
-    public function vigenteDesde(): int;
-
-    public function vigenteHasta(): int;
 }
