+++ src/VigenciasInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PhpCfdi\SatCatalogos;
+
+interface VigenciasInterface
+{
+    public function vigenteDesde(): int;
+
+    public function vigenteHasta(): int;
+}
